refactor(bot-manager): move queue utils to common
