+++ apps/bot-manager/src/common/utils/default-job-options.ts
@@ -1,0 +1,10 @@
+import { DefaultJobOptions } from 'bullmq';
+
+export const defaultJobOptions: DefaultJobOptions = {
+  attempts: Number.MAX_SAFE_INTEGER,
+  backoff: {
+    type: 'custom',
+  },
+  removeOnComplete: true,
+  removeOnFail: true,
+};
